+++ src/Point/BritishNationalGridPoint.php
@@ -29,15 +29,15 @@ class BritishNationalGridPoint extends ProjectedPoint
 {
     private const GRID_LETTERS = 'VWXYZQRSTULMNOPFGHJKABCDE';
 
-    public function __construct(Length $easting, Length $northing, ?DateTimeInterface $epoch = null)
+    public function __construct(Length $easting, Length $northing, ?DateTimeInterface $epoch = null, ?Length $accuracy = null)
     {
-        parent::__construct(Projected::fromSRID(Projected::EPSG_OSGB36_BRITISH_NATIONAL_GRID), $easting, $northing, null, null, $epoch, null);
+        parent::__construct(Projected::fromSRID(Projected::EPSG_OSGB36_BRITISH_NATIONAL_GRID), $easting, $northing, null, null, $epoch, null, $accuracy);
     }
 
     /**
      * @param string $reference OS grid reference (e.g. "TG514131")
      */
-    public static function fromGridReference(string $reference, ?DateTimeInterface $epoch = null): self
+    public static function fromGridReference(string $reference, ?DateTimeInterface $epoch = null, ?Length $accuracy = null): self
     {
         $reference = str_replace(' ', '', $reference);
 
@@ -61,7 +61,7 @@ class BritishNationalGridPoint extends ProjectedPoint
         $easting = $majorEasting + $minorEasting + ((int) substr($numericPortion, 0, (int) $numericPortionSize) * $gridSizeInMetres);
         $northing = $majorNorthing + $minorNorthing + ((int) substr($numericPortion, -(int) $numericPortionSize, (int) $numericPortionSize) * $gridSizeInMetres);
 
-        return new self(new Metre($easting), new Metre($northing), $epoch);
+        return new self(new Metre($easting), new Metre($northing), $epoch, $accuracy);
     }
 
     /**